Add embedded_components to ComponentSpec
Remove embedded from ComponentSpec.

This new field provides a way to specify which components of the embedded
package are actually present inside the component(s) of the parent package.

It is now expected that an embedded package is defined only once in an
entry in InstallSpec.embedded.

Now that install.components[].embedded is removed, it is no longer possible
to nest embedded packages.

Signed-off-by: J Robert Ray <jrray@jrray.org>